fix: Align integer type with GraphQL spec by using int32

+++ internal/server/graph/profile.go
@@ -8,12 +8,12 @@ import (
 	"dreamkast-weaver/internal/server/graph/model"
 )
 
-func newProfile(confName model.ConfName, profileID int) (application.Profile, error) {
+func newProfile(confName model.ConfName, profileID int32) (application.Profile, error) {
 	var e, err error
 	profile := application.Profile{}
 	profile.ConfName, e = value.NewConfName(value.ConferenceKind(confName))
 	err = errors.Join(err, e)
-	profile.ID, e = value.NewProfileID(int32(profileID))
+	profile.ID, e = value.NewProfileID(profileID)
 	err = errors.Join(err, e)
 
 	return profile, err
